‘aprocess’ and ‘sprocess’ snippets grouped.
Problem description:
The ‘aproc’ (i.e. asynchronous process ) snippet is not correct, since it does
not contain the declaration of an asynchronous process (as states) but the
declaration of an synchronous process initialized by a reset signal.

Solution:
The ‘aproc’ snippet has been gouped with the ‘sproc’ under the umbrella of the
synchronous process.
Now, the ‘sproc’ snippet generates the code associated with the reset signal by
default, but this code is removed if the reset signal is removed from the
process sensitivity list during the snippet customization (via substitutions)