Make the sync process id label optional
The synchronous process may not be labeled (removing the default label).
When the label is not defined, the associated code (i.e. colon) is removed.

The code generated by the snippet now contains the block name as closure label
instead of a trailing comment.
It is supported by the VHDL grammar and unlike the trailing comments, there’s no
risk of become outdated due to a block renaming, which leads to misleading
comments.